+++ fpdf/table.py
@@ -502,8 +502,9 @@ class Table:
                 link=cell.link,
             )
 
+            assert "rendered_height" in image
             img_height = (
-                image.rendered_height
+                image["rendered_height"]
                 + padding.top
                 + padding.bottom
                 + cell_border_line_width